Volume Metadata from Underwood, Distant Horizons, Chap. 2

Metadata from Ted Underwood's reproduction repository for Distant Horizons, chap. 2: publication information, record locators, and genre tags derived from bibliographic sources for a selection of science-fiction, detective, Gothic, and "random" fiction.

Usage

genre_meta

Format

genre_meta

A data frame with 1047 rows and 15 columns, notably

docid

unique identifier corresponding to extracted features filename

firstpub

year of first publication of title

tags

string with genre tags, separated by " | "

Source

https://github.com/tedunderwood/horizon/blob/master/chapter2/metadata/concatenatedmeta.csv; for the description of genre tags, https://github.com/tedunderwood/horizon/blob/master/chapter2/metadata/genrecategorieschapter2.docx; and see Ted Underwood, Distant Horizons: Digital Evidence and Literary Change (Chicago: University of Chicago Press, 2019), chap. 2.

Details

Underwood's information about the tags column:

tag	texts	dates	description or source

det100	89	1829 - 1941	The First Hundred Years of Detective Fiction, 1841-1941. 1973. Lilly Library, Bloomington, IN. http://www.indiana.edu/~liblilly/etexts/detective/

chimyst	146	1923-1989	Works categorized by librarians as “detective” or “mystery fiction,” collected at the Chicago Text Lab.

locdetmyst	45	1832-1922	Works categorized by librarians as “detective and mystery fiction,” collected in HathiTrust.

locdetective	16	1865-1912	Works categorized by librarians with the subject heading “Detectives.” Often casebook fiction.

crime	2	1972-1974	Works categorized by librarians as “crime fiction” but not “detective fiction.”

cozy	10	1920-1952	Works by authors mentioned as writing country-house mysteries in The Mystery Readers' Advisory: The Librarian's Clues to Murder and Mayhem, by John Charles, Joanna Morrison, and Candace Clark (Chicago: ALA, 2002).

hardboiled	10	1929-1970	Appendix to Geoffrey O’Brien, Hardboiled America (New York: Van Nostrand Reinhold, 1981).

newgate	7	1828-1840	Keith Hollingsworth, The Newgate Novel, Detroit 1963.

sensation	14	1860-1880	“The Sensation Novel,” Winifred Hughes, in A Companion to the Victorian Novel, ed. Patrick Brantlinger and William B. Thesing (Blackwell, 2002).

lockandkey	10	1800-1903	Works anthologized in The Lock and Key Library: Classic Mystery and Detective Stories, edited by Julian Hawthorne (New York, 1909). Includes writers like Dostoevsky who probably were not seen as writers of mysteries even in 1909; not included in the article’s model of “detective fiction.”

pbgothic	96	1764-1988	“Chronology” in David Punter and Glennis Byron, The Gothic (Malden: Blackwell, 2004).

stangothic	21	1791-1834	A small subset of works tagged as “Gothic” in Stanford Literary Lab metadata.

lochorror	4	1818	Works tagged as “horror” by librarians, collected in HathiTrust.

chihorror	23	1933-1989	Works tagged as “horror” by librarians, collected in the Chicago Text Lab.

locghost	28	1826-1922	Works tagged as “ghost stories” by librarians.

locscifi	21	1836-1909	Works tagged as “science fiction” by librarians and collected in HathiTrust.

chiscifi	144	1901-1989	Works tagged as “science fiction” by librarians and collected at the Chicago Text Lab.

femscifi	9	1818-1922	Ockerbloom, Mary Mark. 2015. “Pre-1950 Utopias and Science Fiction by Women.”

anatscifi	36	1771-1922	Stableford, Brian. 2004. “The Emergence of Science Fiction” and “Science Fiction Between the Wars.” In Anatomy of Wonder, edited by Neil Barron, 5th edition, 3-44.

chiutopia	13	1920-1976	Works tagged as “Utopias” by librarians, collected at the Chicago Text Lab, not folded into “science fiction” in this article.

chifantasy	53	1901-1989	Works tagged as “fantastic” or “fantasy fiction” by librarians, not folded into “science fiction” for the purposes of this article.

juvenile	23	1904-1922	Works for a juvenile audience; collected but not used in this article.

drop	33	1838-1922	Works that I decided not to use, left in the metadata for transparency. The most common reason is that they are juvenile.

random	169	1769-1922	Works randomly selected from HathiTrust Digital Library, using fiction metadata developed in the NEH-funded project “Understanding Genre in a Collectiono f a Million Volumes.” “Random selection” here means that the volumes were selected randomly but then approved or rejected by the author, to avoid stray volumes of nonfiction, classical poetry, juvenile works, etc.

chirandom	202	1920-1989	Works randomly selected from the Chicago Text Lab. Selection here was more genuinely random. Note that both “random” tags can coexist with other genre tags. A randomly-selected volume could also be “chimyst,” for instance; in that case it will be excluded from the negative (contrast) set only if “chimyst” is in the positive set.

teamred	484	1760-1989	Randomly selected authors for a sanity check.

teamblack	500	1764-1989	Randomly selected authors for a sanity check.

stew	224	1764-1989	A random selection of volumes balanced between Gothic, science fiction, and crime/detective traditions, in order to create a ghastly genre stew.